+++ src/tests/arith.test.mjs
@@ -1,27 +1,28 @@
-import { data, trait, apply, dataDecl, extend } from '../index.mjs'
+import { complect, data, trait, extend } from '../index.mjs'
 
 describe('Arithmetic', () => {
     // data declaration
-    const Exp = data({
+    const expData = data({
         Lit: { value: {} },
         Add: { left: {}, right: {} }
-    }),
-        { Add, Lit } = Exp
+    })
 
     // operations
-    const evaluate = trait(Exp, {
+    const evaluate = trait(expData, {
         Lit({ value }) { return value },
         Add({ left, right }) {
-            return this[apply](left) + this[apply](right)
+            return left.evaluate() + right.evaluate()
         }
     })
 
     test('named parameters', () => {
+        const { Lit, Add } = expData
+
         expect(Lit).toBeDefined()
         expect(Add).toBeDefined()
 
         // 1 + (2 + 3)
-        const exp = Add({
+        const e = Add({
             left: Lit({ value: 1 }),
             right: Add({
                 left: Lit({ value: 2 }),
@@ -29,39 +30,46 @@ describe('Arithmetic', () => {
             })
         })
 
-        expect(exp.left.value).toBe(1)
-        expect(exp.right.left.value).toBe(2)
-        expect(exp.right.right.value).toBe(3)
+        expect(e.left.value).toBe(1)
+        expect(e.right.left.value).toBe(2)
+        expect(e.right.right.value).toBe(3)
     })
 
     test('extra parameters throw', () => {
+        const { Lit, Add } = expData
         expect(() => Add({ left: Lit(1), right: Lit(2), extra: 3 })).toThrow()
     })
 
     test('missing parameters throw', () => {
+        const { Lit, Add } = expData
         expect(() => Add({ left: Lit(1) })).toThrow()
     })
 
     test('positional parameters', () => {
+        const { Lit, Add } = expData
         // 1 + (2 + 3)
-        const exp = Add(
+        const e = Add(
             Lit(1),
             Add(Lit(2), Lit(3))
         )
 
-        expect(exp.left.value).toBe(1)
-        expect(exp.right.left.value).toBe(2)
-        expect(exp.right.right.value).toBe(3)
+        expect(e.left.value).toBe(1)
+        expect(e.right.left.value).toBe(2)
+        expect(e.right.right.value).toBe(3)
     })
 
     test('wrong number of parameters throw', () => {
+        const { Lit, Add } = expData
         expect(() => Add(Lit(1))).toThrow()
         expect(() => Add(Lit(1), Lit(2), Lit(3))).toThrow()
     })
 
     test('evaluate', () => {
+        const exp = complect(expData, { evaluate })
+        const { Lit, Add } = exp
+
         // 1 + (2 + 3)
-        const exp = Add({
+        const e = Add({
             left: Lit({ value: 1 }),
             right: Add({
                 left: Lit({ value: 2 }),
@@ -69,8 +77,7 @@ describe('Arithmetic', () => {
             })
         })
 
-        expect(evaluate[dataDecl]).toBe(Exp)
-        expect(evaluate(exp)).toBe(6)
+        expect(e.evaluate()).toBe(6)
 
         // 1 + (2 + 3) + 4
         // with positional parameters
@@ -81,19 +88,22 @@ describe('Arithmetic', () => {
             Lit(4)
         )
 
-        expect(evaluate(exp2)).toBe(10)
+        expect(exp2.evaluate()).toBe(10)
     })
 
-    const print = trait(Exp, {
+    const print = trait(expData, {
         Lit([value]) { return `${value}` },
         Add([left, right]) {
-            return `${this[apply](left)} + ${this[apply](right)}`
+            return `${left.print()} + ${right.print()}`
         }
     })
 
     test('print', () => {
+        const exp = complect(expData, { print })
+        const { Lit, Add } = exp
+
         // 1 + (2 + 3)
-        const exp = Add({
+        const e = Add({
             left: Lit({ value: 1 }),
             right: Add({
                 left: Lit({ value: 2 }),
@@ -101,100 +111,104 @@ describe('Arithmetic', () => {
             })
         })
 
-        expect(print(exp)).toBe('1 + 2 + 3')
+        expect(e.print()).toBe('1 + 2 + 3')
 
         // 1 + (2 + 3) + 4
         // with positional parameters
-        const exp2 = Add(
+        const e2 = Add(
             Add(Lit(1),
                 Add(Lit(2), Lit(3))
             ),
             Lit(4)
         )
 
-        expect(print(exp2)).toBe('1 + 2 + 3 + 4')
+        expect(e2.print()).toBe('1 + 2 + 3 + 4')
     })
 
-    const MulExp = data({
-        [extend]: Exp,
+    const evalPrint = trait(expData, {
+        Add({ left, right }) {
+            return `${left.print()} + ${right.print()} = ${left.evaluate() + right.evaluate()}`
+        },
+        Lit({ value }) { return `${value}` }
+    })
+
+    test('evalPrint', () => {
+        const exp = complect(expData, { evalPrint, evaluate, print })
+        const { Lit, Add } = exp
+
+        // 1 + (2 + 3)
+        const e = Add(Lit(1), Add(Lit(2), Lit(3)))
+
+        expect(e.evalPrint()).toBe('1 + 2 + 3 = 6')
+    })
+
+    const mulExpData = data({
+        [extend]: expData,
         Mul: { left: {}, right: {} }
     })
 
     test('MulExp data', () => {
-        expect(MulExp.Lit).toBeDefined()
-        expect(MulExp.Add).toBeDefined()
-        expect(MulExp.Mul).toBeDefined()
+        const { Lit, Add, Mul } = mulExpData
+
+        expect(Lit).toBeDefined()
+        expect(Add).toBeDefined()
+        expect(Mul).toBeDefined()
 
         // 1 + (2 * 3)
-        const exp = MulExp.Add({
-            left: MulExp.Lit(1),
-            right: MulExp.Mul({
-                left: MulExp.Lit(2),
-                right: MulExp.Lit(3)
-            })
-        })
+        const e = Add(Lit(1), Mul(Lit(2), Lit(3)))
 
-        expect(exp.left.value).toBe(1)
-        expect(exp.right.left.value).toBe(2)
-        expect(exp.right.right.value).toBe(3)
+        expect(e.left.value).toBe(1)
+        expect(e.right.left.value).toBe(2)
+        expect(e.right.right.value).toBe(3)
     })
 
-    const evalMul = trait(MulExp, {
+    const evalMul = trait(mulExpData, {
         [extend]: evaluate,
-        Mul({ left, right }) { return this[apply](left) * this[apply](right) }
+        Mul({ left, right }) { return left.evaluate() * right.evaluate() }
     })
 
     test('evalMul', () => {
-        // 1 + (2 * 3)
-        const exp = MulExp.Add({
-            left: MulExp.Lit({ value: 1 }),
-            right: MulExp.Mul({
-                left: MulExp.Lit({ value: 2 }),
-                right: MulExp.Lit({ value: 3 })
-            })
-        })
+        const exp = complect(mulExpData, { evaluate: evalMul }),
+            { Lit, Add, Mul } = exp
 
-        expect(evalMul(exp)).toBe(7)
+        // 1 + (2 * 3)
+        const e = Add(Lit(1), Mul(Lit(2), Lit(3)))
+
+        expect(e.evaluate()).toBe(7)
     })
 
-    const printMul = trait(MulExp, {
+    const printMul = trait(mulExpData, {
         [extend]: print,
-        Mul([left, right]) { return `${this[apply](left)} * ${this[apply](right)}` }
+        Mul([left, right]) { return `${left.print()} * ${right.print()}` }
     })
 
     test('printMul', () => {
-        // 1 + (2 * 3)
-        const exp = MulExp.Add({
-            left: MulExp.Lit({ value: 1 }),
-            right: MulExp.Mul({
-                left: MulExp.Lit({ value: 2 }),
-                right: MulExp.Lit({ value: 3 })
-            })
-        })
+        const exp = complect(mulExpData, { print: printMul }),
+            { Lit, Add, Mul } = exp
 
-        expect(printMul(exp)).toBe('1 + 2 * 3')
+        // 1 + (2 * 3)
+        const e = Add(Lit(1), Mul(Lit(2), Lit(3)))
+
+        expect(e.print()).toBe('1 + 2 * 3')
     })
 
-    const isValue = trait(MulExp, {
-        Lit({ value }) { return true },
-        Add({ left, right }) { return false },
-        Mul({ left, right }) { return false }
+    const isValue = trait(mulExpData, {
+        Lit() { return true },
+        Add() { return false },
+        Mul() { return false }
     })
 
     test('isValue', () => {
-        // 1 + (2 * 3)
-        const exp = MulExp.Add({
-            left: MulExp.Lit({ value: 1 }),
-            right: MulExp.Mul({
-                left: MulExp.Lit({ value: 2 }),
-                right: MulExp.Lit({ value: 3 })
-            })
-        })
+        const exp = complect(mulExpData, { isValue }),
+            { Lit, Add, Mul } = exp
 
-        expect(isValue(exp)).toBe(false)
-        expect(isValue(exp.left)).toBe(true)
-        expect(isValue(exp.right)).toBe(false)
-        expect(isValue(exp.right.left)).toBe(true)
-        expect(isValue(exp.right.right)).toBe(true)
+        // 1 + (2 * 3)
+        const e = Add(Lit(1), Mul(Lit(2), Lit(3)))
+
+        expect(e.isValue()).toBe(false)
+        expect(e.left.isValue()).toBe(true)
+        expect(e.right.isValue()).toBe(false)
+        expect(e.right.left.isValue()).toBe(true)
+        expect(e.right.right.isValue()).toBe(true)
     })
 })